Reorganize components and remove more stuff
