--- src/lib/components/Spacer.svelte
@@ -1,1 +1,0 @@
-<div class="bg-zinc-50 h-6 border-y border-zinc-200" />
