+++ src/routes/resume/Workplace.svelte
@@ -1,6 +1,4 @@
 <script lang="ts">
-  import Link from "./Link.svelte";
-
   export let title: string;
   export let company: string;
   export let url: string;
@@ -10,7 +8,7 @@
 
 <div>
   <h3 class="text-xl font-serif mb-0.5">
-    {title} at <Link to={url}>{company}</Link>
+    {title} at <a class="link" href={url}>{company}</a>
   </h3>
   <p class="font-serif mb-2">
     <span class="whitespace-nowrap">{dates}</span>
